move check file contents test
