+++ specparam/tests/io/test_models.py
@@ -3,6 +3,7 @@
 import os
 
 from specparam.core.items import OBJ_DESC
+from specparam.io.files import load_json
 
 from specparam.tests.tsettings import TEST_DATA_PATH
 
@@ -156,3 +157,23 @@ def test_load_event():
 
     tevent = load_event('test_event_all', TEST_DATA_PATH)
     assert tevent
+
+def test_load_file_contents():
+    """Check that loaded files contain the contents they should.
+    Note that if this test fails, it likely stems from an issue from saving.
+    """
+
+    file_name = 'test_model_all'
+    loaded_data = load_json(file_name, TEST_DATA_PATH)
+
+    # Check settings
+    for setting in OBJ_DESC['settings']:
+        assert setting in loaded_data.keys()
+
+    # Check results
+    for result in OBJ_DESC['results']:
+        assert result in loaded_data.keys()
+
+    # Check results
+    for datum in OBJ_DESC['data']:
+        assert datum in loaded_data.keys()
